MOVE: Event to slatekit.common
