+++ src/lib/kotlin/slatekit-common/src/main/kotlin/slatekit/common/Event.kt
@@ -1,4 +1,4 @@
-package slatekit.tracking
+package slatekit.common
 
 import slatekit.common.Identity
 import slatekit.common.log.Logger
@@ -52,30 +52,4 @@ data class Event(
         @JvmField val target: String,
         @JvmField val tag: String,
         @JvmField val fields: List<Triple<String, String, String>>?
-) {
-
-    companion object {
-
-        fun log(logger:Logger, id: Identity, event: Event){
-            val extra = event.fields?.fold("") { acc, info -> acc + ", ${info.first}=${info.second}" }
-            when(event.status) {
-                is Passed.Succeeded  -> logger.info ("id=${id.id}, area=${event.area}, name=${event.name}, uuid=${event.uuid}, success=true , code=${event.status.code}, desc=${event.desc} $extra")
-                is Passed.Pending    -> logger.info ("id=${id.id}, area=${event.area}, name=${event.name}, uuid=${event.uuid}, success=true , code=${event.status.code}, desc=${event.desc} $extra")
-                is Failed.Ignored    -> logger.info ("id=${id.id}, area=${event.area}, name=${event.name}, uuid=${event.uuid}, success=false, code=${event.status.code}, desc=${event.desc} $extra")
-                is Failed.Invalid    -> logger.error("id=${id.id}, area=${event.area}, name=${event.name}, uuid=${event.uuid}, success=false, code=${event.status.code}, desc=${event.desc} $extra")
-                is Failed.Denied     -> logger.error("id=${id.id}, area=${event.area}, name=${event.name}, uuid=${event.uuid}, success=false, code=${event.status.code}, desc=${event.desc} $extra")
-                is Failed.Errored    -> logger.error("id=${id.id}, area=${event.area}, name=${event.name}, uuid=${event.uuid}, success=false, code=${event.status.code}, desc=${event.desc} $extra")
-                is Failed.Unexpected -> logger.error("id=${id.id}, area=${event.area}, name=${event.name}, uuid=${event.uuid}, success=false, code=${event.status.code}, desc=${event.desc} $extra")
-                else                 -> logger.error("id=${id.id}, area=${event.area}, name=${event.name}, uuid=${event.uuid}, success=false, code=${event.status.code}, desc=${event.desc} $extra")
-
-            }
-        }
-
-
-        fun logger(logger: Logger, id: Identity): (Event) -> Unit {
-            return { event: Event ->
-                log(logger, id, event)
-            }
-        }
-    }
-}
+)
