+++ src/lib/kotlin/slatekit-tracking/src/main/kotlin/slatekit/tracking/Events.kt
@@ -1,5 +1,6 @@
 package slatekit.tracking
 
+import slatekit.common.Event
 import slatekit.results.*
 
 /**
